refactor(shared): split git resolution out of source_state (300-LOC compliance)
`shared/scripts/source_state.py` landed at 372 lines in #448 — a real crossing of the
300-line guideline that the anti-ratchet CI gate cannot see (it only guards files
already in the baseline). #449 split the test module along the shape-vs-git seam; this
applies the same cut to the source, because holding the tests to a stricter standard
than the code they cover is not a standard.

* `source_state.py` (267) — the identifier's SHAPE: the model, the markdown banner
  form, the JSON block form, and the value validators.
* `source_state_git.py` (123) — what the state actually WAS: HEAD plus whether tracked
  files were modified, with every git-unavailability path degrading to None.

Two importers follow the move (`stamp_test_results.py`, `test_source_state_git.py`);
the compliance renderers and the Group E normaliser are untouched, since `banner_line`,
`SourceState` and `strip_banner` all stay in `source_state.py`.

Behaviour-preserving: 110 cases pass unchanged, plus the full compliance suite (1404)
and both cross-plugin file-path loaders (security 6, shared 5) that the ADR-045
collision would have broken. Verified end-to-end by running the real tool.

Also carries the triage close-out for `trg-4d5b6a56`, delivered by #448.

Run-ID: iterate-2026-07-27-artifact-state-stamping
