+++ .shipwright/triage.jsonl
@@ -887,3 +887,4 @@
 {"event":"status","id":"trg-b14f96d9","ts":"2026-07-27T08:16:54.974770Z","newStatus":"dismissed","by":"complianceBacklog","reason":"complianceRefreshed","promotedTaskId":null}
 {"event":"append","id":"trg-62316d55","ts":"2026-07-27T08:16:54.982555Z","originalTs":"2026-07-27T08:16:54.982555Z","source":"compliance","severity":"high","kind":"compliance","title":"Compliance: 3 open finding(s)","detail":"3 open compliance finding(s): D/D1, D/D3, H/H1\n\n- D/D1: Spec FR coverage in events — uncovered FRs — Must: FR-01.17; Should: FR-01.18 | hint: /shipwright-iterate --type change \"reconcile D1 (Spec FR coverage in events) — see .shipwright/compliance/audit-report.md\"\n- D/D3: Promised FRs delivered — FRs minted with no test totals recorded and never named in affected_frs since — FR-01.17, FR-01.18 | hint: /shipwright-iterate --type change \"reconcile D3 (Promised FRs delivered) — see .shipwright/compliance/audit-report.md\"\n- H/H1: Bloat drift (oversize file not in baseline) — shared/scripts/tools/record_review_pass.py (395 > 300); shared/tests/test_audit_e2e_integration.py (322 > 300); shared/tests/test_record_review_pass_cli.py (496 > 300); shared/tests/test_review_findings.py (403 > 300) | hint: /shipwright-iterate --type change \"add files to shipwright_bloat_baseline.json or split them — see Group H1\"\n\nLive view: .shipwright/compliance/dashboard.md","evidencePath":null,"runId":"unknown","commit":"124b5b01bff68f67b917028bd24e0aae5ed89b9a","dedupKey":"compliance:backlog:a7325e88c249","launchPayload":"/shipwright-compliance\n\nContext: 3 open compliance finding(s): D/D1, D/D3, H/H1.\nDashboard: .shipwright/compliance/dashboard.md\nEach finding + hint is listed in this item's detail.","frId":null,"suiteId":null,"eventId":null,"status":"triage","suggestedPriority":"P1","suggestedDomain":"compliance"}
 {"event":"status","id":"trg-b14f96d9","ts":"2026-07-27T08:16:55.044048Z","newStatus":"dismissed","by":"complianceBacklog","reason":"complianceRefreshed","promotedTaskId":null}
+{"event":"status","id":"trg-4d5b6a56","ts":"2026-07-27T09:31:38.633780Z","newStatus":"dismissed","by":"cli","reason":"delivered by PR #448 (iterate-2026-07-27-artifact-state-stamping): one mechanism in shared/scripts/source_state.py used at both call sites — stamp_test_results.py writes a source_state block into the test-results record; compliance lib/_provenance.py gives all 5 evidence renderers a Source-State line off the SAME work event as Generated:. Identifier = the run id (a SHA cannot exist pre-F6). Stamping only; the enforcing gates stay with trg-12b4cf3f and trg-a1fd8125.","promotedTaskId":null}
